programe/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.18
langchain-openai==0.2.8
llama-index==0.11.23
llama-index-agent-openai==0.3.4
llama-index-embeddings-openai==0.2.5
llama-index-cli==0.3.1
llama-index-core==0.11.23
llama-index-readers-file==0.3.0
llama-index-readers-llama-parse==0.3.0
llama-index-indices-managed-llama-cloud==0.4.2
llama-index-legacy==0.9.48.post4
llama-index-llms-openai==0.2.16
llama-index-multi-modal-llms-openai==0.2.3
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-parse==0.5.14
python-dotenv==1.0.1
dataclasses-json==0.6.7
Jinja2==3.1.2
requests==2.32.3
openai==1.54.4
PyPDF2==3.0.1
pandas==2.2.3
numpy==1.26.4
regex==2024.11.6
pytz==2024.2
typing-extensions==4.12.2
